[chore][pkg/stanza] Cleanup journald input operator files
