+++ pkg/stanza/operator/input/journald/config_all.go
