+++ pkg/stanza/operator/input/journald/config_nonlinux.go
